Simplify component inti further
