+++ components/peripherals/peripherals/environment/Ds18B20SoilSensor.hpp
@@ -64,12 +64,12 @@ public:
         : PeripheralFactory<SinglePinDeviceConfig, EmptyConfiguration>("environment:ds18b20", "environment") {
     }
 
-    std::shared_ptr<Peripheral<EmptyConfiguration>> createPeripheral(const std::string& name, const std::shared_ptr<SinglePinDeviceConfig> deviceConfig, std::shared_ptr<MqttRoot>  /*mqttRoot*/, const PeripheralServices& services) override {
+    std::shared_ptr<Peripheral<EmptyConfiguration>> createPeripheral(const std::string& name, const std::shared_ptr<SinglePinDeviceConfig>& deviceConfig, const std::shared_ptr<MqttRoot>& /*mqttRoot*/, const PeripheralServices& services) override {
         auto sensor = std::make_shared<Ds18B20SoilSensor>(deviceConfig->pin.get());
         services.telemetryCollector->registerProvider("temperature", name, [sensor](JsonObject& telemetryJson) {
             telemetryJson["value"] = sensor->getTemperature();
         });
-        return std::make_shared<SimplePeripheral<Ds18B20SoilSensor>>(name, sensor);
+        return std::make_shared<SimplePeripheral>(name, sensor);
     }
 };
