+++ components/peripherals/peripherals/flow_control/FlowControl.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <memory>
+#include <utility>
 
 #include <Configuration.hpp>
 #include <mqtt/MqttDriver.hpp>
@@ -8,9 +9,7 @@
 #include <peripherals/Peripheral.hpp>
 #include <peripherals/flow_meter/FlowMeter.hpp>
 #include <peripherals/valve/Valve.hpp>
-#include <peripherals/valve/ValveComponent.hpp>
 #include <peripherals/valve/ValveConfig.hpp>
-#include <utility>
 
 using namespace farmhub::kernel::mqtt;
 using namespace farmhub::peripherals;
@@ -29,29 +28,24 @@ class FlowControl : public Peripheral<FlowControlConfig> {
 public:
     FlowControl(
         const std::string& name,
-        const std::shared_ptr<MqttRoot>& mqttRoot,
-        const std::shared_ptr<PulseCounterManager>& pulseCounterManager,
-        const std::shared_ptr<TelemetryPublisher>& telemetryPublisher,
-        std::unique_ptr<ValveControlStrategy> strategy,
-        const InternalPinPtr& pin,
-        double qFactor,
-        milliseconds measurementFrequency)
+        const std::shared_ptr<Valve>& valve,
+        const std::shared_ptr<FlowMeter>& flowMeter)
         : Peripheral<FlowControlConfig>(name)
-        , valve(name, std::move(strategy), mqttRoot, telemetryPublisher)
-        , flowMeter(name, pulseCounterManager, pin, qFactor, measurementFrequency) {
+        , valve(valve)
+        , flowMeter(flowMeter) {
     }
 
     void configure(const std::shared_ptr<FlowControlConfig> config) override {
-        valve.setSchedules(config->schedule.get());
+        valve->setSchedules(config->schedule.get());
     }
 
     void shutdown(const ShutdownParameters /*parameters*/) override {
-        valve.closeBeforeShutdown();
+        valve->closeBeforeShutdown();
     }
 
 private:
-    ValveComponent valve;
-    FlowMeter flowMeter;
+    std::shared_ptr<Valve> valve;
+    std::shared_ptr<FlowMeter> flowMeter;
 
     friend class FlowControlFactory;
 };
@@ -79,30 +73,31 @@ public:
         , Motorized(motors) {
     }
 
-    std::shared_ptr<Peripheral<FlowControlConfig>> createPeripheral(const std::string& name, const std::shared_ptr<FlowControlDeviceConfig> deviceConfig, std::shared_ptr<MqttRoot> mqttRoot, const PeripheralServices& services) override {
+    std::shared_ptr<Peripheral<FlowControlConfig>> createPeripheral(const std::string& name, const std::shared_ptr<FlowControlDeviceConfig>& deviceConfig, const std::shared_ptr<MqttRoot>& mqttRoot, const PeripheralServices& services) override {
         auto strategy = deviceConfig->valve.get()->createValveControlStrategy(this);
 
-        auto flowMeterConfig = deviceConfig->flowMeter.get();
-        auto peripheral = std::make_shared<FlowControl>(
+        auto valve = std::make_shared<Valve>(
             name,
-            mqttRoot,
-            services.pulseCounterManager,
-            services.telemetryPublisher,
-
             std::move(strategy),
+            mqttRoot,
+            services.telemetryPublisher);
 
+        auto flowMeterConfig = deviceConfig->flowMeter.get();
+        auto flowMeter = std::make_shared<FlowMeter>(
+            name,
+            services.pulseCounterManager,
             flowMeterConfig->pin.get(),
             flowMeterConfig->qFactor.get(),
             flowMeterConfig->measurementFrequency.get());
 
-        services.telemetryCollector->registerProvider("flow", name, [peripheral](JsonObject& telemetry) {
-            peripheral->flowMeter.populateTelemetry(telemetry);
+        services.telemetryCollector->registerProvider("valve", name, [valve](JsonObject& telemetry) {
+            valve->populateTelemetry(telemetry);
         });
-        services.telemetryCollector->registerProvider("valve", name, [peripheral](JsonObject& telemetry) {
-            peripheral->valve.populateTelemetry(telemetry);
+        services.telemetryCollector->registerProvider("flow", name, [flowMeter](JsonObject& telemetry) {
+            flowMeter->populateTelemetry(telemetry);
         });
 
-        return peripheral;
+        return std::make_shared<FlowControl>(name, valve, flowMeter);
     }
 };
