+++ components/peripherals/peripherals/flow_meter/FlowMeter.hpp
@@ -103,12 +103,12 @@ public:
         : PeripheralFactory<FlowMeterDeviceConfig, EmptyConfiguration>("flow-meter") {
     }
 
-    std::shared_ptr<Peripheral<EmptyConfiguration>> createPeripheral(const std::string& name, const std::shared_ptr<FlowMeterDeviceConfig> deviceConfig, std::shared_ptr<MqttRoot> /*mqttRoot*/, const PeripheralServices& services) override {
+    std::shared_ptr<Peripheral<EmptyConfiguration>> createPeripheral(const std::string& name, const std::shared_ptr<FlowMeterDeviceConfig>& deviceConfig, const std::shared_ptr<MqttRoot>& /*mqttRoot*/, const PeripheralServices& services) override {
         auto meter = std::make_shared<FlowMeter>(name, services.pulseCounterManager, deviceConfig->pin.get(), deviceConfig->qFactor.get(), deviceConfig->measurementFrequency.get());
         services.telemetryCollector->registerProvider("flow", name, [meter](JsonObject& telemetry) {
             meter->populateTelemetry(telemetry);
         });
-        return std::make_shared<SimplePeripheral<FlowMeter>>(name, meter);
+        return std::make_shared<SimplePeripheral>(name, meter);
     }
 };
